HMRC-657: Handle defaulting to v1 and never matching v2
