+++ config/initializers/rabl_config.rb
@@ -1,4 +1,9 @@
 require 'rabl'
+require 'v1_api'
+
+Rabl::Engine.class_eval do
+  include V1Api.routes.url_helpers
+end
 
 Rabl.configure do |config|
   config.include_json_root = false
